Move RecordReaderFactory and FileFormat to spi
